Cherry-pick upstream r237929
Itanium mangler: don't trip an assertion when unresolved members have implicit bases.

R=sbc@chromium.org

When we find a member of the current instantation, the base of the
unresolved member expression is implicit; use nullptr for such
bases. This is not a change in behavior: the AST already contains null
in such cases, so non-asserts builds do the right thing already. Fixes
rdar://problem/21020559.
Review-Url: https://codereview.chromium.org/2954043003 .